(update) moving all movies related routes under the /movies routes
