+++ src/routes/index.ts
@@ -4,7 +4,7 @@ const router = express.Router();
 /* GET home page. */
 router.get('/', function (_req, res, _next) {
   
-  // #swagger.tags = ['/']
+  // #swagger.tags = ['Base']
   
   res.render('index', { title: 'Welcome to the unofficial Canal Olympia API' });
 });
