+++ src/routes/theaters.ts
@@ -43,74 +43,4 @@ router.get(
   });
 
 
-
-router.get(
-  '/movies/:theater',
-
-  param('theater').isString(),
-  query('lang').default('fr').isIn(['fr', 'en']),
-
-  ExpressValidatorMiddleware,
-
-  async function (req, res, _next) {
-
-    // #swagger.tags = ['Theaters']
-
-    const params = {
-      theaterName: req.params.theater,
-      lang: req.query.lang as string,
-    };
-
-    const response = await theatersController.getMovies(params, res);
-
-    return response;
-  });
-
-
-router.get(
-  '/movie-infos/:slug',
-
-  param('slug').isString(),
-  query('lang').default('fr').isIn(['fr', 'en']),
-
-  ExpressValidatorMiddleware,
-
-  async function (req, res, _next) {
-
-    // #swagger.tags = ['Theaters']
-
-    const params = {
-      slug: req.params.slug,
-      lang: req.query.lang as string,
-    };
-
-    const response = await theatersController.getMovieInfoBySlug(params, res);
-
-    return response;
-  });
-
-
-router.get(
-  '/movie-diffusion-infos/:slug',
-
-  param('slug').isString(),
-  query('lang').default('fr').isIn(['fr', 'en']),
-
-  ExpressValidatorMiddleware,
-
-  async function (req, res, _next) {
-
-    // #swagger.tags = ['Theaters']
-
-    const params = {
-      slug: req.params.slug,
-      lang: req.query.lang as string,
-    };
-
-    const response = await theatersController.getMovieDiffusionInfos(params, res);
-
-    return response;
-  });
-
-
 export default router;
